Moved restart policy related functionality to a separate file.
